iss-98 refactor auth
